Loop approx and different bases support
Finished idea for solving the looping problem. Works fine for some cases. Same for change of basis. However, some of the algebras are really glitchy, and this behavior also results from changing the basis on non glitchy algs. Not sure if it's a basis thing with LoopApprox, or inherit to the alg's parameters.